clk: ipq40xx: minor updates to adcc clk driver
some mask bits had to be updated. new clocks frequencies had
been requested. the updates are for these frequencies

+++ drivers/clk/qcom/clk-rcg.h
@@ -20,7 +20,7 @@
 struct freq_tbl {
 	unsigned long freq;
 	u8 src;
-	u8 pre_div;
+	u16 pre_div;
 	u16 m;
 	u16 n;
 };
